Create a full test and add comments in defaults
This commit create the entire test structure that can be used by
Travis and also by any user who wants to immediately test the role.
It also updates the default/main.yml file so to be more clear about
how it is possible to change things to get what a user wants.